non-ROUTINE! patches for enabling -Wsign-compare
Signed and unsigned comparison rules are kind of annoying, but they
are important to know.  Quick quiz: what does this output?

    #include <stdio.h>
    int main() {
        unsigned int e = 304;
        int b = -1020;
        if (e + b > 1020) puts(">1020"); else puts("<=1020");
    }

The sum will be driven to be an unsigned value, since one addends
was unsigned.  And an unsigned value can't hold a negative number.
So it's good to be aware of these.  Sometimes the answer is that
you didn't need to mix after all, sometimes it just needs a cast
to make it explicit and call attention.

This is all the changes that weren't in ROUTINE!/STRUCT! so they
could be considered separately.

+++ src/core/f-dtoa.c
@@ -556,7 +556,8 @@ Balloc
 	int x;
 	Bigint *rv;
 #ifndef Omit_Private_Memory
-	unsigned int len;
+	size_t len; // !!! Ren/C: (unsigned int => size_t) for -Wsign-compare
+	// !!! REVIEW: isn't size_t unsigned?  Why did that fix it?
 #endif
 
 	ACQUIRE_DTOA_LOCK(0);
@@ -3567,9 +3568,9 @@ strtod
 
  static char *
 #ifdef KR_headers
-rv_alloc(i) int i;
+rv_alloc(i) size_t i; // !!! Ren/C: (int => size_t) for -Wsign-compare
 #else
-rv_alloc(int i)
+rv_alloc(size_t i) // !!! Ren/C: (int => size_t) for -Wsign-compare
 #endif
 {
 	int j, k, *r;
